% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PQSViews-class.R
\docType{methods}
\name{as.character,PQSViews-method}
\alias{as.character,PQSViews-method}
\title{Coerce to character vector}
\usage{
\S4method{as.character}{PQSViews}(x)
}
\arguments{
\item{x}{PQSViews object.}
}
\value{
Character vector representing PQS.
}
\description{
Coerce to character vector
}

