% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PQSViews-class.R
\docType{methods}
\name{density,PQSViews-method}
\alias{density,PQSViews-method}
\title{Get density vector}
\usage{
\S4method{density}{PQSViews}(x)
}
\arguments{
\item{x}{PQSViews object}
}
\value{
Density vector
}
\description{
Desity vector represents numbers of PQS (potential quadruplex forming
sequences) overlapping at each position in input sequence.
}
\examples{
pqs <- pqsfinder(DNAString("CCCCCCGGGTGGGTGGGTGGGAAAA"))
density(pqs)

}

