% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PQSViews-class.R
\docType{methods}
\name{maxScores,PQSViews-method}
\alias{maxScores,PQSViews-method}
\title{Get vector of maximal scores}
\usage{
\S4method{maxScores}{PQSViews}(x)
}
\arguments{
\item{x}{PQSViews object.}
}
\value{
Vector of maximal scores.
}
\description{
For each sequence position it gives the maximal score
of all PQS conformations which overlap that position.
}
\examples{
pqs <- pqsfinder(DNAString("CCCCCCGGGTGGGTGGGTGGGAAAA"))
maxScores(pqs)

}
