% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PQSViews-class.R
\name{maxScores}
\alias{maxScores}
\title{Get vector of maximal scores}
\usage{
maxScores(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Additional arguments, for use in specific methods.}
}
\value{
Vector of maximal scores.
}
\description{
Get vector of maximal scores for a given object.
}
\examples{
showMethods("maxScores")

}
