% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pqsfinder}
\alias{pqsfinder}
\title{Identificate potential quadruplex forming sequences.}
\usage{
pqsfinder(subject, max_len = 70L, run_min_len = 3L, run_max_len = 11L,
  loop_min_len = 0L, loop_max_len = 30L, g_bonus = 20L,
  bulge_penalty = 10L, run_re = "G{1,5}.{0,5}G{1,5}", user_fn = NULL,
  use_cache = 1L, use_re = 0L, use_prof = 0L, debug = 0L)
}
\arguments{
\item{subject}{DNAString object.}

\item{max_len}{Maximal lenth of PQS.}

\item{run_min_len}{Minimal length of quadruplex run.}

\item{run_max_len}{Maximal length of quadruplex run.}

\item{loop_min_len}{Minimal length of quadruplex loop.}

\item{loop_max_len}{Maxmimal length of quadruplex loop.}

\item{g_bonus}{Score bonus for one complete G tetrade.}

\item{bulge_penalty}{Penalization for a bulge in quadruplex run.}

\item{run_re}{Regular expression specifying one run of quadruplex.}

\item{user_fn}{Custom quadruplex scoring function. It takes the following 10
arguments: \code{subject} - Input DNAString object, \code{score} - implicit PQS score,
\code{start} - PQS start position, \code{width} - PQS width, \code{loop_1} - start pos. of loop #1,
\code{run_2} - start pos. of run #2, \code{loop_2} - start pos. of loop #2,
\code{run_3} - start pos. of run #3, \code{loop_3} - start pos. of loop #3,
\code{run_4} - start pos. of run #4. Return value of the function should be new score
represented as asingle integer value.}

\item{use_cache}{Use cache for low complexity regions?}

\item{use_re}{Use regular expression engine to validate quadruplex run?}

\item{use_prof}{Enables profiling.}

\item{debug}{Enables detailed debugging output. Turn it on if you want
to see all possible quadruplexes found at each positions and not just
the best one.}
}
\value{
\code{\link{PQSViews}} object
}
\description{
Function for identification of all potential intramolecular quadruplex
patterns (PQS) in DNA sequence.
}
\examples{
pv <- pqsfinder(DNAString("CCCCCCGGGTGGGTGGGTGGGAAAA"))

}

