% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PQSViews-class.R
\docType{class}
\name{PQSViews-class}
\alias{.PQSViews}
\alias{PQSViews-class}
\title{An S4 class to represent potential quadruplex forming sequences.}
\description{
An S4 class to represent potential quadruplex forming sequences.
}
\section{Slots}{

\describe{
\item{\code{density}}{Numbers of PQS (potential quadruplex forming sequences)
starting at each position in input sequence.}
}}

