% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PQSViews-class.R
\name{PQSViews}
\alias{PQSViews}
\title{PQSViews class constructor.}
\usage{
PQSViews(subject, start, end, score, density)
}
\arguments{
\item{subject}{DNAString object.}

\item{start}{Vector of PQS start positions.}

\item{end}{Vector of PQS end positions.}

\item{score}{Vector of PQS scores.}

\item{density}{Numbers of PQS starting at each position in \code{subject}.}
}
\value{
PQSViews object
}
\description{
User friendly constructor of PQSViews class representing potential
quadruplex forming sequences (PQS).
}
\examples{
pv <- PQSViews(DNAString("CGGGCGGGGC"), 1:2, 2:3, 10:11, 1:10)

}

