% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PQSViews-class.R
\docType{methods}
\name{score,PQSViews-method}
\alias{score,PQSViews-method}
\title{Get PQS score vector}
\usage{
\S4method{score}{PQSViews}(x)
}
\arguments{
\item{x}{PQSViews object}
}
\value{
Score vector
}
\description{
Get PQS score vector
}
\examples{
pqs <- pqsfinder(DNAString("CCCCCCGGGTGGGTGGGTGGGAAAA"))
score(pqs)

}

