% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PQSViews-class.R
\docType{methods}
\name{scoreDistribution,PQSViews-method}
\alias{scoreDistribution,PQSViews-method}
\title{Get score distribution vector}
\usage{
\S4method{scoreDistribution}{PQSViews}(x)
}
\arguments{
\item{x}{PQSViews object.}
}
\value{
Score distribution vector.
}
\description{
Score distribution vector represents score of the best PQS found
at each sequence position.
}
\examples{
pqs <- pqsfinder(DNAString("CCCCCCGGGTGGGTGGGTGGGAAAA"))
scoreDistribution(pqs)

}

