% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PQSViews-class.R
\name{scoreDistribution}
\alias{scoreDistribution}
\title{Get score distribution vector}
\usage{
scoreDistribution(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Additional arguments, for use in specific methods.}
}
\value{
Score distribution vector.
}
\description{
Get score distribution vector of given object.
}
\examples{
showMethods("scoreDistribution")

}

