% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PQSViews-class.R
\docType{methods}
\name{show,PQSViews-method}
\alias{show,PQSViews-method}
\title{Show method}
\usage{
\S4method{show}{PQSViews}(object)
}
\arguments{
\item{object}{PQSViews object.}
}
\value{
PQSViews object printed.
}
\description{
Show method
}

