% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pqsfinder}
\alias{pqsfinder}
\title{Identificate potential quadruplex forming sequences.}
\usage{
pqsfinder(subject, max_len = 70L, min_score = 0L, run_min_len = 3L,
  run_max_len = 11L, loop_min_len = 0L, loop_max_len = 30L,
  tetrad_bonus = 20L, bulge_penalty = 10L, mismatch_penalty = 10L,
  run_re = "G{1,5}.{0,5}G{1,5}", custom_scoring_fn = NULL,
  use_default_scoring = TRUE, verbose = FALSE)
}
\arguments{
\item{subject}{DNAString object.}

\item{max_len}{Maximal lenth of PQS.}

\item{min_score}{Minimal PQS score.}

\item{run_min_len}{Minimal length of quadruplex run.}

\item{run_max_len}{Maximal length of quadruplex run.}

\item{loop_min_len}{Minimal length of quadruplex loop.}

\item{loop_max_len}{Maxmimal length of quadruplex loop.}

\item{tetrad_bonus}{Score bonus for one complete G tetrade.}

\item{bulge_penalty}{Penalization for a bulge in quadruplex run.}

\item{mismatch_penalty}{Penalization for a mismatch in tetrad.}

\item{run_re}{Regular expression specifying one run of quadruplex.}

\item{custom_scoring_fn}{Custom quadruplex scoring function. It takes the
following 10 arguments: \code{subject} - Input DNAString object,
\code{score} - implicit PQS score, \code{start} - PQS start position,
\code{width} - PQS width, \code{loop_1} - start pos. of loop #1,
\code{run_2} - start pos. of run #2, \code{loop_2} - start pos. of loop
#2, \code{run_3} - start pos. of run #3, \code{loop_3} - start pos. of
loop #3, \code{run_4} - start pos. of run #4. Return value of the function
has to be new score represented as a single integer value. Please note
that if \code{use_default_scoring} is enabled, the custom scoring function
is evaluated AFTER the default scoring system but ONLY IF the default
scoring system resulted in non-zero score (for performance reasons). On
the other hand, when \code{use_default_scoring} is disabled, custom
scoring function is evaluated on every PQS.}

\item{use_default_scoring}{Enables default internal scoring system. This
option is particularly useful in case you intend to radically change the
default behavior and specify your own scoring function. By disabling the
default scoring you will get a full control above the underlying detection
algorithm.}

\item{verbose}{Enables detailed output. Turn it on if you want to see all
possible PQS found at each positions and not just the best one. It is
highly recommended to use this option for debugging custom quadruplex
scoring function. Each PQS is reported on separate row in the following
format: \code{start cnt pqs_sequence score}, where \code{start} is the PQS
starting position, \code{pqs_sequence} shows the PQS sequence structure
with each run surrounded by square brackets and \code{score} is the score
assigned to the particular PQS by all applied scoring functions.}
}
\value{
\code{\link{PQSViews}} object
}
\description{
Function for identification of all potential intramolecular quadruplex
patterns (PQS) in DNA sequence.
}
\examples{
pv <- pqsfinder(DNAString("CCCCCCGGGTGGGTGGGTGGGAAAA"))

}

