% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PQSViews-class.R
\name{PQSViews}
\alias{PQSViews}
\title{PQSViews class constructor}
\usage{
PQSViews(subject, start, width, strand, score, density, max_scores, nt, nb, nm,
  ll1, ll2, ll3)
}
\arguments{
\item{subject}{DNAString object.}

\item{start}{Start positions.}

\item{width}{Lengths.}

\item{strand}{Strand specifications.}

\item{score}{Scores.}

\item{density}{Numbers of PQS overlapping at each position in \code{subject}.}

\item{max_scores}{Score of the best PQS found at each position.}

\item{nt}{Tetrad numbers.}

\item{nb}{Bulge counts.}

\item{nm}{Mismatch counts.}

\item{ll1}{Loop 1 lengths.}

\item{ll2}{Loop 2 lengths.}

\item{ll3}{Loop 3 lengths.}
}
\value{
PQSViews object.
}
\description{
User friendly constructor for PQSViews class representing potential
quadruplex forming sequences (PQS). PQSViews is a subclass of
\code{\link{XStringViews}} class and adds two more slots to store
PQS density and PQS score distribution.
}
\details{
Use \code{\link{elementMetadata}} function to get extra PQS features
like number of tetrads, bulges, mismatches or loop lengths.
}
\examples{
pv <- PQSViews(DNAString("CGGGCGGGGC"), c(1,2), c(2,3), c("+", "+"),
               c(10,11), 1:10, 1:10, 0, 0, 0, 1, 1, 1)
start(pv)
width(pv)
strand(pv)
score(pv)
density(pv)
maxScores(pv)
elementMetadata(pv)

}

