% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PQSViews-class.R
\name{PQSViews}
\alias{PQSViews}
\title{PQSViews class constructor.}
\usage{
PQSViews(subject, start, width, strand, score, density)
}
\arguments{
\item{subject}{DNAString object.}

\item{start}{Vector of PQS start positions.}

\item{width}{Vector of PQS lengths.}

\item{strand}{Vector of PQS strand specifications.}

\item{score}{Vector of PQS scores.}

\item{density}{Numbers of PQS overlapping at each position in \code{subject}.}
}
\value{
PQSViews object
}
\description{
User friendly constructor for PQSViews class representing potential
quadruplex forming sequences (PQS). PQSViews is a subclass of
\code{\link{XStringViews}} class and adds one more slot to store
PQS density.
}
\examples{
pv <- PQSViews(DNAString("CGGGCGGGGC"), 1:2, 2:3, "+", 10:11, 1:10)
start(pv)
width(pv)
strand(pv)
score(pv)
density(pv)

}

